fix(expressions): the legacy dialect unifies an integral with a float as a float (#299)
Spark's legacy dialect takes the tightest common type for an integral against a
FLOAT, which is float, and rounds the integral onto it first. ANSI widens the
pair to double. We answered ANSI's rule under both. Measured on 4.0.3 at every
integral width, the split reaches every site that unifies, not only the
comparisons the issue named: =, <, <=>, IN, BETWEEN, simple CASE, coalesce,
if, CASE, nvl2, greatest, least, and arithmetic + - * %. `/` stays double, and
a decimal or double against a float is double in both dialects.

- SparkNumericTypes.CommonType and ArithmeticResult take the dialect.
- ComparisonTarget and SetComparisonTarget answer float under legacy, and the
  evaluator's MightRound gate now asks about a float as well as a decimal.
- A bigint reaches a float in ONE rounding, as Java's (float) long does:
  through double, 2^60 + 2^36 + 1 ties and lands on 2^60 instead of
  2^60 + 2^37. This also fixes CAST(bigint AS FLOAT) in both dialects. The
  string and decimal sources have the same flaw and are #372.

Measured into a new `integral-float` corpus group under both dialects, gated on
type as well as value; the seven #299 known differences are gone.

+++ src/EngineeredWood.Expressions.Arrow/Spark/SparkFunctions.cs
@@ -382,8 +382,8 @@ internal static class SparkFunctions
             var floats = new FloatArray.Builder();
             for (var i = 0; i < rowCount; i++)
             {
-                var value = choice[i] < 0 ? null : SparkArrays.ReadDouble(sources[choice[i]], i);
-                if (value is null) floats.AppendNull(); else floats.Append((float)value.Value);
+                var value = choice[i] < 0 ? null : SparkArrays.ReadFloat(sources[choice[i]], i);
+                if (value is null) floats.AppendNull(); else floats.Append(value.Value);
             }
 
             return floats.Build();
